fix: properly wait for docs engine shutdown (#2389)
## Description

We did not properly await the shutdown of the docs actor. This was
likely the cause for a flaky test `default_author_persist`, because it
might happen that the node shutdown ended without the docs store being
dropped.

This PR fixes this by properly awaiting for the reply from the docs
actor, signalling completion of the shutdown, thus making sure that the
docs redb is dropped before `node.shutdown()` terminates.

Reenables the `default_author_perist` test :crossed_fingers: .

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ]~~ Documentation updates if relevant.~~
- [ ] ~~Tests if relevant.~~
- [x] All breaking changes documented.

Co-authored-by: Rüdiger Klaehn <rklaehn@protonmail.com>