chore: clippy fixes (#54)
## Description

* Fixes for clippy 1.89
* Quite a few methods in the ranger::Store trait are unused outside of
tests. This fixes this partially. The trait was used more before we made
the redb store the only store (and the "memory" store now is the "fs"
store with the in-memory redb backend). This should be revisited
properly, i.e. actually remove the things that are not needed anymore.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.